% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Affichesudoko.R
\name{AfficheSudoko}
\alias{AfficheSudoko}
\title{Affichage d'un sudoku}
\usage{
AfficheSudoko(z, Couleur)
}
\arguments{
\item{z}{: Une matrice}

\item{Couleur}{: une matrice qui contient des 0 et des 1}
}
\value{
: Plot de la grille du sudoku
}
\description{
Affichage d'un sudoku
}
\examples{

 A = Suduko()
 A= genererSoduko(A,50)
 Couleur = matrix (0,9,9)
 AfficheSudoko(A,Couleur)


}
\author{
AIT MOULAY ABDERRAHIM
}
