% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BacktrackingFinale.R
\name{BacktrackingFinale}
\alias{BacktrackingFinale}
\title{Résoudre un Sudoku grace a l'algorithme du backtraking}
\usage{
BacktrackingFinale(B)
}
\arguments{
\item{B}{: Une matrice}
}
\value{
une Data frame qui contient la solution du grille de sudoku avec ainsi que le nombre de back.
}
\description{
Résoudre un Sudoku grace a l'algorithme du backtraking
}
\examples{
 A = Suduko()
 A = genererSoduko(A,50)
 BacktrackingFinale(A)


}
\author{
AIT MOULAY ABDERRAHIM
}
