% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genererSoduko.R
\name{genererSoduko}
\alias{genererSoduko}
\title{Suprimer des valeur aliatoire dans un Soduku}
\usage{
genererSoduko(A, N)
}
\arguments{
\item{A}{: une matrice}

\item{N}{; nombre de valeur a suprimer}
}
\value{
une matrice de tail 9x9 qui contient les valeurs vide
}
\description{
Suprimer des valeur aliatoire dans un Soduku
}
\examples{
 A = matrix(1,9,9)
A = genererSoduko (A , 54 )
}
\author{
AIT MOULAY ABDERRAHIM
}
