% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenererSoduko_avec_difficulte.R
\name{GenererSudoku_avec_difficulte}
\alias{GenererSudoku_avec_difficulte}
\title{Generer un Sudoku avec difficulte grace a l'algorithme du backtraking}
\usage{
GenererSudoku_avec_difficulte(Niveau)
}
\arguments{
\item{Niveau}{: un entier entre 1 et 4}
}
\value{
une Data frame qui contient la grille du sudoku avec son solution ansi que le nombre de back.
}
\description{
Generer un Sudoku avec difficulte grace a l'algorithme du backtraking
}
\examples{

 GenererSudoku_avec_difficulte(2)


}
\author{
AIT MOULAY ABDERRAHIM
}
