% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitBlock.R
\name{GitBlock}
\alias{GitBlock}
\title{La fonction qui return le num?ro du block}
\usage{
GitBlock(i, j)
}
\arguments{
\item{i}{: l'indice du ligne}

\item{j}{: l'indice du colonne}
}
\value{
: Num?ro du block
}
\description{
La fonction qui return le num?ro du block
}
\examples{
GitBlock(2 , 3)
}
\author{
AIT MOULAY ABDERRAHIM
}
