% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid}
\alias{is_valid}
\title{Test une valeur si il est valide dans une case de la matrice}
\usage{
is_valid(A, i, j, n)
}
\arguments{
\item{A}{: une matrice}

\item{i}{: l'indice du ligne}

\item{j}{: l'indice du colonne}

\item{n}{; valeur rechercher}
}
\value{
: True si elle est pas valide, Sinon False
}
\description{
Test une valeur si il est valide dans une case de la matrice
}
\examples{
 A = Suduko()
 A= genererSoduko(A,50)
 is_valid(A , 3, 2 , 3 )

}
\author{
AIT MOULAY ABDERRAHIM
}
