% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Possibilite.R
\name{Possiblite}
\alias{Possiblite}
\title{Possiblite
Data frame des indices des valeurs vides ainsi que leur nombre de possibilte}
\usage{
Possiblite(A)
}
\arguments{
\item{A}{: une matrice}
}
\value{
: Data Frame
}
\description{
Possiblite
Data frame des indices des valeurs vides ainsi que leur nombre de possibilte
}
\examples{
A = Suduko()
A= genererSoduko(A,50)

Possiblite (A)
}
\author{
AIT MOULAY ABDERRAHIM
}
