% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suduko.R
\name{Suduko}
\alias{Suduko}
\title{Generation d'un suduko}
\usage{
Suduko()
}
\arguments{
\item{:}{Sans paramétre}
}
\value{
une matrice de tail 9x9 qui contient les chiffres du suduko
}
\description{
Generation d'un suduko
}
\examples{
A = Suduko()
}
\author{
AIT MOULAY ABDERRAHIM
}
