% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testBlock.R
\name{testBlock}
\alias{testBlock}
\title{Test une valeur si il existe dans un block}
\usage{
testBlock(A, i, j, n)
}
\arguments{
\item{A}{: une matrice}

\item{i}{: l'indice du ligne}

\item{j}{: l'indice du colonne}

\item{n}{; valeur rechercher}
}
\value{
: True si n existe deja dans le bock, Sinon False
}
\description{
Test une valeur si il existe dans un block
}
\examples{
 A = matrix(1,9,9)
testBlock (A , 2 , 5 , 3 )
}
\author{
AIT MOULAY ABDERRAHIM
}
