% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcolonne.R
\name{testcolonne}
\alias{testcolonne}
\title{Test une valeur si il existe dans une colonne}
\usage{
testcolonne(A, j, n)
}
\arguments{
\item{A}{: une matrice}

\item{j}{: l'indice du colonne}

\item{n}{; valeur rechercher}
}
\value{
: True si n existe deja dans la colonne, Sinon False
}
\description{
Test une valeur si il existe dans une colonne
}
\examples{
A = Suduko()
A= genererSoduko(A,50)
testcolonne (A , 2 , 3 )
}
\author{
AIT MOULAY ABDERRAHIM
}
