% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testligne.R
\name{testligne}
\alias{testligne}
\title{Test une valeur si il existe dans une ligne}
\usage{
testligne(A, i, n)
}
\arguments{
\item{A}{: une matrice}

\item{i}{: l'indice du ligne}

\item{n}{; valeur rechercher}
}
\value{
: True si n existe deja dans la ligne, Sinon False
}
\description{
Test une valeur si il existe dans une ligne
}
\examples{
A = matrix(1,9,9)
testligne (A , 2 , 5 )
}
\author{
AIT MOULAY ABDERRAHIM
}
