% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSuduko.R
\name{TestSuduko}
\alias{TestSuduko}
\title{Test la solution du Suduko si elle est valide ou non}
\usage{
TestSuduko(A)
}
\arguments{
\item{A}{: une matrice}
}
\value{
: True si la solution est valide, sinon False
}
\description{
Test la solution du Suduko si elle est valide ou non
}
\examples{
 A = Suduko()
 A= genererSoduko(A,50)
 TestSuduko (A)
}
\author{
AIT MOULAY ABDERRAHIM
}
